Langchain/1.ChatModels/chatmodel_google.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
import os

load_dotenv()

api_key=os.getenv('GOOGLE_API_KEY')
model=ChatGoogleGenerativeAI(model='gemini-2.5-flash',temperature=1,api_key=api_key)
result=model.invoke('Who is Narendra Modi?')
print(result.content)